glyph_info {systemfonts}	R Documentation

Query glyph-specific information from fonts

Description

This function allows you to extract information about the individual glyphs in a font, based on a specified size. All size related measures are in pixel-units. The function is vectorised to the length of the glyphs vector.

Usage

glyph_info(
  glyphs,
  family = "",
  italic = FALSE,
  bold = FALSE,
  size = 12,
  res = 72,
  path = NULL,
  index = 0
)

Arguments

glyphs
A vector of glyphs. Strings will be split into separate glyphs automatically

family
The name of the font family

italic
logicals indicating the font style

bold
logicals indicating the font style

size
The pointsize of the font to use for size related measures

res
The ppi of the size related mesures

path, index
path an index of a font file to circumvent lookup based on family and style

Value

A data.frame with information about each glyph, containing the following columns:

glyph

The glyph as a character

index

The index of the glyph in the font file

width

The width of the glyph

height

The height of the glyph

x_bearing

The horizontal distance from the origin to the leftmost part of the glyph

y_bearing

The vertical distance from the origin to the top part of the glyph

x_advance

The horizontal distance to move the cursor after adding the glyph

y_advance

The vertical distance to move the cursor after adding the glyph

bbox

The tight bounding box surrounding the glyph

[Package systemfonts version 1.0.1 Index]