string_width {systemfonts}	R Documentation

Calculate the width of a string, ignoring new-lines

Description

This is a very simple alternative to shape_string() that simply calculates the width of strings without taking any newline into account. As such it is suitable to calculate the width of words or lines that has already been splitted by ⁠\n⁠. Input is recycled to the length of strings.

Usage

string_width(
  strings,
  family = "",
  italic = FALSE,
  bold = FALSE,
  size = 12,
  res = 72,
  include_bearing = TRUE,
  path = NULL,
  index = 0
)

Arguments

strings
A character vector of strings

family
The name of the font family

italic
logicals indicating the font style

bold
logicals indicating the font style

size
The pointsize of the font to use for size related measures

res
The ppi of the size related mesures

include_bearing
Logical, should left and right bearing be included in the string width?

path
path an index of a font file to circumvent lookup based on family and style

index
path an index of a font file to circumvent lookup based on family and style

Value

A numeric vector giving the width of the strings in pixels. Use the provided res value to convert it into absolute values.

Examples

strings <- c('A short string', 'A very very looong string')
string_width(strings)

[Package systemfonts version 1.0.5 Index]