systemfonts-package {systemfonts}	R Documentation

systemfonts: System Native Font Finding

Description

Provides system native access to the font catalogue. As font handling varies between systems it is difficult to correctly locate installed fonts across different operating systems. The 'systemfonts' package provides bindings to the native libraries on Windows, macOS and Linux for finding font files that can then be used further by e.g. graphic devices. The main use is intended to be from compiled code but 'systemfonts' also provides access from R.

Author(s)

Maintainer: Thomas Lin Pedersen thomas.pedersen@rstudio.com (ORCID)

Authors:

Jeroen Ooms jeroen@berkeley.edu (ORCID)

Devon Govett (Author of font-manager)

Other contributors:

RStudio [copyright holder]

See Also

Useful links:

https://github.com/r-lib/systemfonts

Report bugs at https://github.com/r-lib/systemfonts/issues

[Package systemfonts version 1.0.1 Index]